+++ rut_chile/rut_chile.py
@@ -15,11 +15,9 @@ def is_valid_rut(rut: str) -> bool:
         bool -- True if rut is valid. False otherwise
     """
 
-    format_regex = r"^((\d{1,3}(\.\d{3})+-)|\d+-?)(\d|k|K)$"  # Valid format
-    if not rut or not re.match(format_regex, rut):
+    if not rut or not __is_well_formatted(rut):
         return False
-
-    rut = rut.replace(".", "").replace("-", "").lower()  # Standardize input
+    rut = __clean_rut(rut)
     return get_verification_digit(rut[:-1]) == rut[-1]
 
 
@@ -32,15 +30,14 @@ def get_verification_digit(rut: str, capitalize: bool = False) -> str:
         exception
 
         capitalize {bool} -- Indicates if returned value must be a capital
-        letter. False by default
+        letter (default: {False})
 
     Returns:
         str -- Verification digit. It might be a digit, 'k' or 'K'.
     """
     format_regex = r"^\d+$"  # Regex to check valid format
     if not rut or not re.match(format_regex, rut):
-        raise Exception("Invalid input")
-
+        return None
     factors = [2, 3, 4, 5, 6, 7]  # Factors for calculating verification digit
     partial_sum = 0
     factor_pos = 0
@@ -54,3 +51,62 @@ def get_verification_digit(rut: str, capitalize: bool = False) -> str:
             return "K"
         return "k"
     return str(verification_digit)
+
+
+def format_rut(rut: str, with_dots: bool = False, upper: bool = True) -> str:
+    """Formats RUT according to the options
+
+    Arguments:
+        rut {str} -- RUT to be formatted
+
+    Keyword Arguments:
+        with_dots {bool} -- Indicates whether to add dot separators
+        (default: {False})
+        upper {bool} -- Indicates if result should be in uppercase
+        (default: {True})
+
+    Returns:
+        str -- Formatted RUT. If input is not valid, returns None
+    """
+
+    if not rut or not __is_well_formatted(rut):
+        return None
+    rut = __clean_rut(rut)
+
+    if with_dots:  # Add dots
+        formatted_rut = __add_thousands_separator(rut[:-1])
+    else:
+        formatted_rut = rut[:-1]
+
+    formatted_rut = '-'.join([formatted_rut, rut[-1]])  # Add dash
+
+    if upper:
+        formatted_rut = formatted_rut.upper()
+
+    return formatted_rut
+
+
+def __is_well_formatted(rut: str) -> bool:
+    if not rut:
+        raise ValueError("rut cannot be None")
+    format_regex = r"^((\d{1,3}(\.\d{3})+-)|\d+-?)(\d|k|K)$"  # Valid format
+    return re.match(format_regex, rut) is not None
+
+
+def __clean_rut(rut: str) -> bool:
+    if not rut or not __is_well_formatted(rut):
+        raise ValueError("rut must be well formatted")
+    return rut.replace(".", "").replace("-", "").lower()  # Standardize input
+
+
+def __add_thousands_separator(rut: str) -> str:
+    if len(rut) < 4:
+        return rut
+
+    digit_groups = []
+    if len(rut) % 3 > 0:  # Most significant group of digits
+        digit_groups.append(rut[:len(rut) % 3])
+    for i in range(int(len(rut) / 3)):  # Generate 3-digit groups
+        start = len(rut) % 3 + 3 * i
+        digit_groups.append(rut[start:start + 3])
+    return ".".join(digit_groups)